Created en/de(crypt) directories & test file
Directories for encrypt and decrypt data was created to store data
to encrypt and to store decrypted data. A test file was created to
test out the functionality of FileEncryptor.java and the tests ran
successfully. Files from the resources directory was encrypted and
stored into the encrypted_data directory and then decrypted and
final result was stored in the decrypted_data directory. gitignore
file was also updated to ignore any class and xml plugin files.