Created Server.java and set up Server class
A server file was created that prints to the terminal that it is
active and listening on a specific port. It then, checks the system
if a client file storing directory has been created or not and
creates it depending on the scenario. The server then accepts clients
and receives any files sent from the client and stores it in the
received_files directory.