Created Client.java & Edited Server.java
Created Client.java so any client can send their choice of file over
to the server. The client connects to the server at the specified
port and using the file within the resources directory, the client
transfers the file over to the server. Since this program will be
run outside of my local device, I had to implement relative pathing
so Server.java saw a change in the pathing. Overall, files from the
client are being transferred over to the server.