Implemented encryption when sending  from client
Implemented the encryptFile method from FileEncryptor.java to
encrypt a file when sending to the server. Edited FileEncryptor.java
to turn it into a package. Edited Client.java to allow for terminal
input for files to transfer to the server as well as to input a
password which gets used for the AES encryption.